tests: refactor terraform tests to not depend on each other
